+++ gradle/dependencies.gradle
@@ -20,7 +20,7 @@ ext.dep = [
     supportAnnotations      : 'com.android.support:support-annotations:25.1.0',
     compiletesting          : 'com.google.testing.compile:compile-testing:0.10',
     javaPoet                : 'com.squareup:javapoet:1.8.0',
-    moshi                   : 'com.squareup.moshi:moshi:1.3.1',
+    moshi                   : 'com.squareup.moshi:moshi:1.4.0',
     rxjava                  : 'io.reactivex:rxjava:1.2.3',
     rxjava2                 : 'io.reactivex.rxjava2:rxjava:2.0.5',
     rxandroid               : 'io.reactivex.rxjava2:rxandroid:2.0.1',
